1、修改过时依赖：dataBinding 2、修改副屏中视频声音控制的方法 3、在副屏中添加视频全屏和自适应宽高的显示、以及是否播放声音的控制 4、修改MQTT文件下载方法 5、添加限制：机器必须对接充电桩才能下发任务 6、修改讲解队列中切换任务点混乱问题 7、SeekBar可为负数的问题 8、添加测温和人脸识别限制：前提是有相对应的硬件设备

+++ app/src/main/java/com/sendi/deliveredrobot/model/BasicModel.java
@@ -9,8 +9,8 @@ public class BasicModel {
     private String robotMode;//机器人音色
     private Boolean Intelligent;//智能语音
     private Boolean Etiquette;//礼仪迎宾
-    private float voiceVolume;//语音音量
-    private float videoVolume; //视频音量
+    private int voiceVolume;//语音音量
+    private int videoVolume; //视频音量
     private Boolean expression;//操作屏表情
     private Boolean identifyVip;//识别Vip人脸迎宾
     private float leadingSpeed;//引领速度
@@ -68,19 +68,19 @@ public class BasicModel {
         Etiquette = etiquette;
     }
 
-    public float getVoiceVolume() {
+    public int getVoiceVolume() {
         return voiceVolume;
     }
 
-    public void setVoiceVolume(float voiceVolume) {
+    public void setVoiceVolume(int voiceVolume) {
         this.voiceVolume = voiceVolume;
     }
 
-    public float getVideoVolume() {
+    public int getVideoVolume() {
         return videoVolume;
     }
 
-    public void setVideoVolume(float videoVolume) {
+    public void setVideoVolume(int videoVolume) {
         this.videoVolume = videoVolume;
     }
